move KeyStoreUtil to public API
